Extract globalAgg tests into its own file
